README.md/requirement.txt
pytesseract
pdf2image
numpy
scikit-learn
Pillow
